Fix ordering when no title is available
We were originally using the title to sort values, this only worked if the titles were unique.

If the user did not choose to include titles, or had duplicate titles, then this would break the logic meaning only the first value with that title would be shown (multiple times).

We've fixed that by adding an index to the stored values array, which we then use to sort after the fact.